run_on_service: handle unexpected command exit
The command would only be launched if a matching service was added. If
the command crashed it would create a zombie process that would not be
reaped and relaunched until the matching service was restarted. This
patch fixes that. The zeroconf thread simply updates the availability
and information of the matching service. The main thread launches the
command in an infinite loop contingent on the availability of the
matching service.

Perhaps in the future 'run' can return if the failure rate becomes
excessive.